Changed screen color depth to 32bpp.

+++ src/game/Laptop/AIMMembers.h
@@ -11,7 +11,7 @@ void HandleAIMMembers(void);
 void RenderAIMMembers(void);
 
 
-void DrawNumeralsToScreen(INT32 iNumber, INT8 bWidth, UINT16 usLocX, UINT16 usLocY, SGPFont, UINT8 ubColor);
+void DrawNumeralsToScreen(INT32 iNumber, INT8 bWidth, UINT16 usLocX, UINT16 usLocY, SGPFont, UINT32 ubColor);
 
 void DisplayTextForMercFaceVideoPopUp(const ST::string& str);
 void EnterInitAimMembers(void);
