+++ src/game/Laptop/IMP_Attribute_Finish.cc
@@ -67,8 +67,8 @@ static void MakeButton(UINT idx, const ST::string& text, INT16 y, GUI_CALLBACK c
 {
 	BUTTON_PICS* const img = LoadButtonImage(LAPTOPDIR "/button_2.sti", 0, 1);
 	giIMPAttributeFinishButtonImage[idx] = img;
-	const INT16 text_col   = FONT_WHITE;
-	const INT16 shadow_col = DEFAULT_SHADOW;
+	const UINT32 text_col   = FONT_WHITE;
+	const UINT32 shadow_col = DEFAULT_SHADOW;
 	GUIButtonRef const btn = CreateIconAndTextButton(img, text, FONT12ARIAL, text_col, shadow_col, text_col, shadow_col, LAPTOP_SCREEN_UL_X + 130, y, MSYS_PRIORITY_HIGH, click);
 	giIMPAttributeFinishButton[idx] = btn;
 	btn->SetCursor(CURSOR_WWW);
