+++ src/game/Laptop/IMP_MainPage.cc
@@ -127,8 +127,8 @@ static void MakeButton(UINT idx, const char* img_file, const ST::string& text, I
 {
 	BUTTON_PICS* const img = LoadButtonImage(img_file, 0, 1);
 	giIMPMainPageButtonImage[idx] = img;
-	const INT16 text_col   = FONT_WHITE;
-	const INT16 shadow_col = DEFAULT_SHADOW;
+	const UINT32 text_col   = FONT_WHITE;
+	const UINT32 shadow_col = DEFAULT_SHADOW;
 	GUIButtonRef const btn = CreateIconAndTextButton(img, text, FONT12ARIAL, text_col, shadow_col, text_col, shadow_col, x, y, MSYS_PRIORITY_HIGH, click);
 	giIMPMainPageButton[idx] = btn;
 	btn->SetCursor(CURSOR_WWW);
