+++ src/game/Tactical/Interface_Utils.h
@@ -11,7 +11,7 @@
 
 void DrawSoldierUIBars(SOLDIERTYPE const&, INT16 sXPos, INT16 sYPos, BOOLEAN fErase, SGPVSurface* buffer);
 
-void DrawItemUIBarEx(OBJECTTYPE const&, UINT8 ubStatus, INT16 sXPos, INT16 sYPos, INT16 sHeight, INT16 sColor1, INT16 sColor2, SGPVSurface* buffer);
+void DrawItemUIBarEx(OBJECTTYPE const&, UINT8 ubStatus, INT16 sXPos, INT16 sYPos, INT16 sHeight, UINT32 sColor1, UINT32 sColor2, SGPVSurface* buffer);
 
 void RenderSoldierFace(SOLDIERTYPE const&, INT16 sFaceX, INT16 sFaceY);
