+++ src/sgp/Shading.h
@@ -7,9 +7,9 @@ void BuildShadeTable(void);
 void BuildIntensityTable(void);
 void SetShadeTablePercent(float uiShadePercent);
 
-extern UINT16 IntensityTable[65536];
-extern UINT16 ShadeTable[65536];
-extern UINT16 White16BPPPalette[256];
+extern UINT32 IntensityTable[65536];
+extern UINT32 ShadeTable[65536];
+extern UINT32 White16BPPPalette[256];
 
 #define DEFAULT_SHADE_LEVEL 4
